Image Convert

+++ CMI/Engine/Asset/ParameterSettings/ViewerConversionSettings.cs
@@ -9,7 +9,7 @@ namespace CMI.Engine.Asset.ParameterSettings
         [Description("Gibt vor, welche Qualität die eingefügten Bilder haben. Durch eine tiefere Qualität erreicht man eine kleinere Dateigrösse.")]
         [DefaultValue(80)]
         [Validation("^[1-9][0-9]?$|^100$", "Geben Sie eine Ganzzahl grösser 0 und kleiner gleich 100 ein.")]
-        public long JpegQualitaetInProzent { get; set; }
+        public int JpegQualitaetInProzent { get; set; }
 
         [Description("Default-Auflösung für resultierende JPEG, in dpi")]
         [DefaultValue(96)]
